fix(remote): stop orphaning billable VMs on swallowed cleanup failure (#870)
`VMPoolManager::cleanup_idle_vms` removed each idle VM from the pool and then
discarded the deallocation result with `let _ = ... .cleanup(...)`. When
`Orchestrator::cleanup` failed, the VM was silently dropped from tracking while
still running in the cloud — an ongoing, undetected cost leak.

Inspect the `Result<bool, RemoteError>` structurally (never by parsing tool
output) and branch exhaustively via a new pure helper
`VMPoolManager::apply_cleanup_result`:

- `Ok(true)`  — deallocation confirmed: drop the VM and record it in `removed`.
- `Ok(false)` — ran but unconfirmed: retain the entry, `warn!`, retry next pass.
- `Err(e)`    — hard failure: retain the entry, `error!`, never orphan the VM.

`removed` now reflects only truthfully reclaimed VMs, so the `save_state` guard
and the batch `info!` count stay honest. The `Err` arm logs only the curated
`RemoteError` Display and the VM name — never raw tool stdout/stderr, tags, or
credentials.

Tests: add branch-coverage unit tests for all three outcomes. The test module
is moved to a sibling `vm_pool_tests.rs` (wired via `#[path]`, still a child
module with private-item access) so `vm_pool.rs` stays within the issue #536
<=500-line module budget enforced by `remote_rust_modules_stay_under_500_lines`.

Also fix two pre-existing workspace clippy errors that block the mandatory
pre-commit/CI `cargo clippy -- -D warnings` gate under the current stable
toolchain (redundant borrow in amplihack-blarify grep; `match` -> `?` in
amplihack-hooks cwd), so the gate is green without `--no-verify`.

Closes #870
