+++ crates/amplihack-hooks/src/pre_tool_use/cwd/mod.rs
@@ -158,16 +158,16 @@ fn check_glob_cwd_match(pattern: &str, cwd: &Path) -> Option<Value> {
     }
 
     let prefix_path = Path::new(prefix);
-    let glob_dir = match prefix_path.parent() {
-        Some(d) => match d.canonicalize() {
+    let glob_dir = {
+        let d = prefix_path.parent()?;
+        match d.canonicalize() {
             Ok(c) => c,
             Err(_) => return None,
-        },
-        None => return None,
+        }
     };
-    let basename_prefix = match prefix_path.file_name() {
-        Some(n) => n.to_string_lossy().to_string(),
-        None => return None,
+    let basename_prefix = {
+        let n = prefix_path.file_name()?;
+        n.to_string_lossy().to_string()
     };
 
     let mut current = Some(cwd.to_path_buf());
